// src/types/partials/ImmunizationProtocolApplied.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TReference } from '../partials/Reference';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TInt } from '../simpleTypes/Int';

export type TImmunizationProtocolApplied = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    series?: String;
    authority?: TReference;
    targetDisease?: TCodeableConcept[];
    doseNumberPositiveInt?: TInt;
    doseNumberString?: String;
    seriesDosesPositiveInt?: TInt;
    seriesDosesString?: String;
};